add place on spline script and update mesh orientation
