+++ Assets/Prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b7f05c069d6612b4f87bb689ed564f72
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
