+++ Assets/Scripts/Runtime/MeshOnSpline.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cb9d1cf5fd124c7469aaf32496156670
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
